Emotion detection (1)/requirements.txt
numpy
pandas
seaborn
neattext
plotly
scikit-learn
joblib
mysql.connector
deep_translator